fixed music volume gauge... actually the bug was mine *blush*, discard the explaination i gave in previous commit
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/branches/irrlicht@3393 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/audio/sound_manager.hpp
@@ -65,7 +65,7 @@ public:
                                                     m_current_music->switchToFastMusic();}
     
     void                    setMasterMusicVolume(float gain);
-    int                     getMasterMusicVolume() const { return m_masterGain; }
+    float                   getMasterMusicVolume() const { return m_masterGain; }
     
     MusicInformation       *getCurrentMusic() {return m_current_music; }    
     MusicInformation       *getMusicInformation(const std::string& filename);
